Persist formatting into database

+++ shared/src/commonMain/kotlin/com/module/notelycompose/notes/domain/DeleteNoteById.kt
@@ -4,7 +4,7 @@ package com.module.notelycompose.notes.domain
 class DeleteNoteById(
     private val noteDataSource: NoteDataSource
 ) {
-    suspend fun execute(id: Int) {
+    suspend fun execute(id: Long) {
         return noteDataSource.deleteNoteById(id)
     }
 }
